% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seffcovar.R
\name{get_ComBat_slide_effect_covariates}
\alias{get_ComBat_slide_effect_covariates}
\title{Constructing slide-effect-adjustment covariates using SVD/Eigenvalue decomposition on ComBat-estimated slide effects}
\usage{
get_ComBat_slide_effect_covariates(
  methylation_matrix,
  slide,
  mod = NULL,
  input_set = S_high
)
}
\arguments{
\item{methylation_matrix}{A matrix with DNA methylation measurements (beta-value or M-scale). Rows correspond to CpG sites and columns to samples. Rownames need to be in 'cgXXXXXX' format.}

\item{slide}{Slide information for each sample in the dataset}

\item{mod}{Covariate information to adjust, default is NULL.}

\item{input_set}{List of slide-effect-susceptible CpG sites from which the covariates will be extracted. Default is S_high.}
}
\value{
pcs Top ten slide effect PCs
}
\description{
This function uses a modification of the ComBat implementation in the sva R package. The estimated gamma.hat (least squares estimate) are used to perform a
singular value decomposition.
The original code for ComBat from the sva package that can be found at
https://bioconductor.org/packages/release/bioc/html/sva.html
The original code as well as this code here is distributed under the Artistic License 2.0.
}
