% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sets.R
\docType{data}
\name{S_0_20}
\alias{S_0_20}
\title{Set of slide-effect-susceptible CpG sites "S_0_20" as derived by Hecker et al. 2023}
\format{
An object of class \code{character} of length 250972.
}
\usage{
data(S_0_20)
}
\description{
This set of slide-effect-susceptible CpG sites was
derived by Hecker et al. 2023. S_0_20 contains all CpG sites excluding S_80_100, S_60_80, S_40_60, and S_20_40.
}
\references{
Hecker et al. (2023)
}
\keyword{datasets}
