% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sets.R
\docType{data}
\name{S_20_40}
\alias{S_20_40}
\title{Set of slide-effect-susceptible CpG sites "S_20_40" as derived by Hecker et al. 2023}
\format{
An object of class \code{character} of length 370770.
}
\usage{
data(S_20_40)
}
\description{
This set of slide-effect-susceptible CpG sites was
derived by Hecker et al. 2023. S_20_40 contains all CpG sites where at least one out of 9
datasets estimated a proportion of >20\% variance explained by slide effects, excluding S_80_100, S_60_80, and S_40_60.
}
\references{
Hecker et al. (2023)
}
\keyword{datasets}
