% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sets.R
\docType{data}
\name{S_40_60}
\alias{S_40_60}
\title{Set of slide-effect-susceptible CpG sites "S_40_60" as derived by Hecker et al. 2023}
\format{
An object of class \code{character} of length 160713.
}
\usage{
data(S_40_60)
}
\description{
This set of slide-effect-susceptible CpG sites was
derived by Hecker et al. 2023. S_40_60 contains all CpG sites where at least one out of 9
datasets estimated a proportion of >40\% variance explained by slide effects, excluding S_80_100 and S_60_80.
}
\references{
Hecker et al. (2023)
}
\keyword{datasets}
