% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sets.R
\docType{data}
\name{S_60_80}
\alias{S_60_80}
\title{Set of slide-effect-susceptible CpG sites "S_60_80" as derived by Hecker et al. 2023}
\format{
An object of class \code{character} of length 33667.
}
\usage{
data(S_60_80)
}
\description{
This set of slide-effect-susceptible CpG sites was
derived by Hecker et al. 2023. S_60_80 contains all CpG sites where at least one out of 9
datasets estimated a proportion of >60\% variance explained by slide effects, excluding S_80_100.
}
\references{
Hecker et al. (2023)
}
\keyword{datasets}
