% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sets.R
\docType{data}
\name{S_80_100}
\alias{S_80_100}
\title{Set of slide-effect-susceptible CpG sites "S_80_100" as derived by Hecker et al. 2023}
\format{
An object of class \code{character} of length 1685.
}
\usage{
data(S_80_100)
}
\description{
This set of slide-effect-susceptible CpG sites was
derived by Hecker et al. 2023. S_80_100 contains all CpG sites where at least one out of 9
datasets estimated a proportion of >80\% variance explained by slide effects.
}
\references{
Hecker et al. (2023)
}
\keyword{datasets}
