% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_sets.R
\docType{data}
\name{S_high}
\alias{S_high}
\title{Set of slide-effect-susceptible CpG sites "S_high" as derived by Hecker et al. 2023}
\format{
An object of class \code{character} of length 1578.
}
\usage{
data(S_high)
}
\description{
This set of slide-effect-susceptible CpG sites was
derived by Hecker et al. 2023. S_high contains all CpG sites where at least 7 out of
9 datasets had an estimated proportion of 60\% variance explained by slide effects.
}
\references{
Hecker et al. (2013)
}
\keyword{datasets}
