% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seffcovar.R
\name{get_slide_effects_covariates}
\alias{get_slide_effects_covariates}
\title{Constructing slide-effect-adjustment covariates}
\usage{
get_slide_effects_covariates(methylation_matrix, scale_parameter = FALSE)
}
\arguments{
\item{methylation_matrix.}{A matrix with DNA methylation measurements (beta-value or M-scale). Rows correspond to CpG sites and columns to samples.}

\item{scale_parameter.}{Scale parameter for the prcomp function, defaults to FALSE.}
}
\description{
This function constructs slide-effect-adjustment covariates from a beta-value or M-value DNA methylation matrix.
}
\examples{
tobedone
}
